+++ assets/js/downloadProgressDialog.js
@@ -39,30 +39,54 @@ const RESULT_TYPE_TO_TOTAL_COUNT_PROPERTY_MAP = {
 
 
 export default class DownloadProgressDialog {
-    constructor(el) {
+    constructor(el, formType) {
         this.el = el;
+        this.formType = formType;
     }
 
+    getFormElements(){
+        let elements;
+        if (this.formType == "advanced"){
+            elements = {'StatusModal': document.getElementById('download-status-modal'), 
+                'DownloadButtons': document.getElementById('downloadButtons'),
+                'Description': document.getElementById('download-modal-description'),
+                'Heading': document.getElementById('download-modal-heading')};
+        }
+        else{
+            elements = {'StatusModal': document.getElementById('download-status-modal-basic'),
+            'DownloadButtons': document.getElementById('downloadButtonsBasic'),
+            'Description': document.getElementById('download-modal-basic-description'),
+            'Heading': document.getElementById('download-status-modal-basic-heading')};
+        }
+        return elements;
+    }
 
-    buttonHtml(id, label) {
-        return '<button id="' + id + '" type="button" class="btn btn-default">' + label + '</button>';
+    buttonHtml(id, text) {
+        return `<li class="usa-button-group__item"><button type="button" class="usa-button" id="${id}" data-close-modal>${text}</button></li>`;
     }
 
     show(thisOpKind, dialogMessage) {
+        let elements = this.getFormElements();
+        if(elements['StatusModal'] && elements['StatusModal'].hidden){
+            elements['StatusModal'].hidden = false;
+        }
         var message = dialogMessage ? dialogMessage : 'Validating query ... Please wait.';
         this.opKind = thisOpKind;
 
-        this.el.find('.modal-footer').html('');
-        this.el.find('.modal-body').html(message);
-        this.el.find('.modal-header h4').html(DIALOG[this.opKind].title);
-        this.el.modal('show');
-    }
-
-    hide() {
-        this.el.modal('hide');
+        elements['DownloadButtons'].innerHTML = '';
+        elements['Description'].innerHTML = message;
+        elements['Heading'].innerHTML = (DIALOG[this.opKind].title);
+        // Remove existing continue and close buttons
+        if (document.getElementById('continueButton')){
+            document.getElementById('continueButton').remove();
+        }
+        if (document.getElementById('closeDownloadModal')){
+            document.getElementById('closeDownloadModal').remove();
+        }
     }
 
     updateProgress(counts, resultType, fileFormat, continueFnc) {
+        let elements = this.getFormElements();
         var totalCount = counts.total[RESULT_TYPE_TO_TOTAL_COUNT_PROPERTY_MAP[resultType]];
 
         var getCountMessage = function () {
@@ -94,24 +118,30 @@ export default class DownloadProgressDialog {
         } else if (DIALOG[this.opKind].cancelDownload(totalCount, fileFormat)) {
             this.cancelProgress(getCountMessage() + DIALOG[this.opKind].cancelMessage);
         } else {
-            this.el.find('.modal-body').html(getCountMessage() + '<p>Click Continue to ' + DIALOG[this.opKind].continueMessage);
-            this.el.find('.modal-footer').html(this.buttonHtml('progress-cancel-btn', 'Cancel') +
-                this.buttonHtml('progress-continue-btn', 'Continue'));
-            $('#progress-cancel-btn').click(() => {
-                this.el.modal('hide');
-            });
-            $('#progress-continue-btn').click(() => {
-                this.el.modal('hide');
+            elements['Description'].innerHTML = `${getCountMessage()}<p>Click Continue to ${DIALOG[this.opKind].continueMessage}`;
+            elements['DownloadButtons'].innerHTML = `${this.buttonHtml('closeDownloadModal', 'Cancel')}${this.buttonHtml('continueButton', 'Continue')}`;
+            document.getElementById('continueButton').onclick = function() {
+                if(elements['StatusModal'] && !elements['StatusModal'].hidden){
+                    elements['StatusModal'].hidden = true;
+                }
                 continueFnc(totalCount);
-            });
+            };
+            document.getElementById('closeDownloadModal').onclick = function() {
+                if(elements['StatusModal'] && !elements['StatusModal'].hidden){
+                    elements['StatusModal'].hidden = true;
+                }
+            };
         }
     }
 
     cancelProgress(message) {
-        this.el.find('.modal-body').html(message);
-        this.el.find('.modal-footer').html(this.buttonHtml('progress-ok-btn', 'Ok'));
-        $('#progress-ok-btn').click(() => {
-            this.el.modal('hide');
-        });
+        let elements = this.getFormElements();
+        elements['Description'].innerHTML = message;
+        elements['DownloadButtons'].innerHTML = this.buttonHtml('progressOkBtn', 'Ok');
+        document.getElementById('progressOkBtn').onclick = function() {
+            if(!elements['StatusModal'].hidden){
+                elements['StatusModal'].hidden = true;
+            }
+        };
     }
 }