+++ assets/js/views/nldiView.js
@@ -154,6 +154,8 @@ export default class NldiView {
                     self.map.addLayer(self.nldiSiteCluster);
 
                     self.updateNldiInput(nldiModel.getUrl('wqp'));
+                    // Update the api query urls
+                    $('#params').trigger('change');
                 })
                 .catch(function (error) {
                     // handle error
@@ -247,6 +249,8 @@ export default class NldiView {
         nldiModel.reset();
         this.cleanUpMaps();
         this.map.closePopup();
+        // Update the api query urls
+        $('#params').trigger('change');
     }
 
     /*
@@ -283,7 +287,7 @@ export default class NldiView {
         const undoImg = Config.STATIC_ENDPOINT;
 
         const searchControl = L.control.searchControl(Config.GEO_SEARCH_API_ENDPOINT);
-        const clearControl = L.easyButton('<img src="' + undoImg + '/img/usa-icons/undo.svg" alt="reset"/><div id="resetText">Reset</div>', this.clearHandler.bind(this), 'Clear the sites', {
+        const clearControl = L.easyButton('<img src="' + undoImg + 'img/usa-icons/undo.svg" alt="reset"/><div id="resetText">Reset</div>', this.clearHandler.bind(this), 'Clear the sites', {
             position: 'topright'
         });